pyproject.toml
[project]
name = "davinci-codex"
version = "1.0.0"
description = "Open, reproducible reconstructions of Leonardo da Vinci's civil inventions."
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "sympy",
    "matplotlib>=3.6",
    "pytest>=7.2",
    "pytest-benchmark>=4.0",
    "ruff>=0.3",
    "mypy>=1.0",
    "typer",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "trimesh",
    "meshio",
    "pillow",
    "rich>=13.0",
]
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Hunter M. Bown", email = "hunter@shannon-labs.com" }
]
keywords = ["leonardo", "davinci", "engineering", "renaissance", "inventions", "simulation", "CAD"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

[project.optional-dependencies]
cad = [
    "pythonocc-core>=7.6; (sys_platform == 'linux' or sys_platform == 'win32') and python_version < '3.11'",
]
docs = [
    "jupyter-book>=1.0",
    "sphinx>=5.0",
    "sphinx-book-theme>=1.0",
    "myst-parser>=0.18",
    "nbclient>=0.7",
    "jupyter-client>=7.0",
    "ipykernel>=6.0",
    "ipywidgets>=8.0",
    "plotly>=5.0",
    "pandas>=1.5",
]

[project.scripts]
davinci-codex = "davinci_codex.cli:app"

[project.urls]
Homepage = "https://github.com/Shannon-Labs/davinci-codex"
Documentation = "https://shannon-labs.github.io/davinci-codex/"
Repository = "https://github.com/Shannon-Labs/davinci-codex.git"
"Bug Tracker" = "https://github.com/Shannon-Labs/davinci-codex/issues"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py39"
line-length = 100
include = [
    "src/**/*.py",
    "tests/**/*.py",
    "synthesis/**/*.py",
]
extend-exclude = [
    "analysis",
    "artifacts",
    "cad",
    "data",
    "docs",
    "education",
    "examples",
    "failures",
    "inventions",
    "materials",
    "mechanical_lion_complete_package",
    "notebooks",
    "src/davinci_codex/visualization",
    "research",
    "sims",
    "tools",
    "validation",
    "vendor",
    "web",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["E501", "B008", "C901", "UP006", "UP007", "UP035", "UP045"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "-q", "--strict-markers"]
testpaths = ["tests"]
pythonpath = "src"
